# s2: Add AMD64 assembly for better mode