… README.md
… s2/README.md
… s2/encode_go.go
… s2/encodeblock_amd64.go
… s2/encodeblock_amd64.s
… s2/s2_test.go